refactor: move caching to its own file
