+++ kernel/src/memory/mmio.rs
@@ -19,8 +19,8 @@
 //! MMIO (Memory-Mapped I/O) allows to access a device's registers by mapping them on the main
 //! memory.
 
-use super::{buddy, PhysAddr, VirtAddr};
-use crate::{arch::x86, memory::vmem::KERNEL_VMEM, process::oom};
+use super::{buddy, oom, PhysAddr, VirtAddr};
+use crate::{arch::x86, memory::vmem::KERNEL_VMEM};
 use core::ptr::NonNull;
 use utils::errno::AllocResult;
