+++ kernel/src/memory/oom.rs
@@ -24,49 +24,32 @@
 //!
 //! This is an emergency procedure which is not supposed to be used under normal conditions.
 
-use crate::sync::mutex::Mutex;
 use utils::errno::AllocResult;
 
-/// The maximum number of times the kernel tries to kill a process to retrieve
-/// memory.
-const MAX_TRIES: u32 = 5;
-
-/// Variable telling whether the OOM killer is enabled.
-static KILLER_ENABLE: Mutex<bool> = Mutex::new(true);
-
-/// Tells whether the OOM killer is enabled.
-pub fn is_killer_enabled() -> bool {
-	*KILLER_ENABLE.lock()
+/// TODO doc
+pub fn reclaim() {
+	// TODO try the following one after the other:
+	// - shrink page cache
+	// - shrink directory entries cache
+	// - swap memory to disk
+	// - if the kernel is configured for it, prompt the user to select processes to kill
+	// - if the kernel is configured for it, kill the process with the highest OOM score (ignore
+	//   init process)
+	// - else, panic:
+	panic!("Out of memory");
 }
 
-/// Enables or disables the OOM killer.
-pub fn set_killer_enabled(enable: bool) {
-	*KILLER_ENABLE.lock() = enable;
-}
-
-/// Runs the OOM killer.
-pub fn kill() {
-	if !is_killer_enabled() {
-		panic!("Out of memory");
-	}
-
-	// TODO Get the process with the highest OOM score (ignore init process)
-}
-
-/// Executes the given function.
-///
-/// On fail due to a lack of memory, the function runs the OOM killer, then tries again.
+/// Executes the given function. On failure due to a lack of memory, the function runs the OOM
+/// killer, then tries again.
 ///
 /// If the OOM killer is unable to free enough memory, the kernel may panic.
 pub fn wrap<T, F: FnMut() -> AllocResult<T>>(mut f: F) -> T {
-	for _ in 0..MAX_TRIES {
+	for _ in 0..5 {
 		if let Ok(r) = f() {
 			return r;
 		}
-
-		kill();
+		reclaim();
 		// TODO Check if current process has been killed
 	}
-
 	panic!("OOM killer is unable to free up space for new allocations!");
 }
